check relevant tx type to only show in related tab
